+++ src/parser/shaman/enhancement/modules/core/Interval.ts
@@ -1,0 +1,29 @@
+const debug = false;
+
+export class Interval {
+  protected startTime: number;
+  protected endTime: number | undefined = undefined;
+
+  constructor(timestamp: number) {
+    this.startTime = timestamp;
+  }
+
+  end(timestamp: number) {
+    this.endTime = timestamp;
+  }
+
+  get duration(): number {
+    if (this.endTime === undefined) {
+      debug &&
+        console.error(
+          'Cannot calculate duration of an Interval with no endTime.',
+        );
+      return 0;
+    }
+    return this.endTime - this.startTime;
+  }
+
+  get ended(): boolean {
+    return this.endTime !== undefined;
+  }
+}
